requirements.txt
google-adk
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy

